Sahar

+++ src/main/java/be/intecbrussel/Service/implemantation/BlogServiceImpl.java
@@ -18,11 +18,11 @@ public class BlogServiceImpl implements BlogService {
 
     @Override
     public List<Blog> getAllBlogs() {
-        return blogsRepository.findAll();
+    return blogsRepository.findAll();
     }
 
     @Override
-    public Blog getBlogById(int id) {
+    public Blog getBlogById(Long id) {
         Blog blog =  blogsRepository.findById(id).get();
         return blog;
     }
