+++ src/main/java/be/intecbrussel/repository/BlogsRepository.java
@@ -4,6 +4,6 @@ package be.intecbrussel.repository;
 import be.intecbrussel.data.Blog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlogsRepository extends JpaRepository<Blog,Integer> {
+public interface BlogsRepository extends JpaRepository<Blog,Long> {
 
 }
